payslips/requirements.txt
# camelot-py[cv]
tabula-py
PyPDF2